refactor(text-sync): split lifecycle helpers (#0000)
